Kavin_Aravindhan_&_Chhavi_Dixit/requirements.txt
coco_lib==0.1.4
httpx==0.28.1
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
Requests==2.32.3
scikit_learn==1.6.1
timm==0.9.12
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1